Generate smaller u64s in random_value_of_type
I'm not entirely sure *why* it's happening (maybe a parse issue causing
overflow) but sometimes with u64 we generate very small *negative*
numbers in INSERT statements, causing them to fail when operating on
mysql. Without digging too much into what's going on there, this just
drops the size down to u32 to avoid the issue.

Reviewed-on: https://gerrit.readyset.name/c/readyset/+/91
Tested-by: Buildkite CI

+++ query_generator/src/lib.rs
@@ -195,7 +195,7 @@ fn random_value_of_type(typ: &SqlType) -> DataType {
         SqlType::Int(_) => rng.gen::<i32>().into(),
         SqlType::Bigint(_) => rng.gen::<i64>().into(),
         SqlType::UnsignedInt(_) => rng.gen::<u32>().into(),
-        SqlType::UnsignedBigint(_) => rng.gen::<u64>().into(),
+        SqlType::UnsignedBigint(_) => rng.gen::<u32>().into(),
         SqlType::Tinyint(_) => rng.gen::<i8>().into(),
         SqlType::UnsignedTinyint(_) => rng.gen::<u8>().into(),
         SqlType::Smallint(_) => rng.gen::<i16>().into(),
